fix: rename the op and change the example
